+++ CHANGELOG.md
@@ -4,6 +4,18 @@ Toutes les modifications notables de ce projet sont documentées dans ce fichier
 
 Si vous souhaitez voir le détail, cliquez sur les numéros qui vous renverrons vers les tickets correspondants.
 
+# [2.13.1] - Correctifs
+## Fix :
+- [#465](https://github.com/Refhi/Weda-Helper/issues/465) - L'historique gauche est de nouveau affiché avec 30% de largeur par défaut au lieu des 50% proposé de base par Weda. (actif si l'historique gauche est activé automatiquement dans les consultations)
+- [#459](https://github.com/Refhi/Weda-Helper/issues/459) - Un bouton permet désormais de réinitialiser le classement automatique des documents dans les échanges sécurisés. Cela permet de tester plus facilement différents mots-clés de classification.
+
+### Imports automatisés :
+- [#448](https://github.com/Refhi/Weda-Helper/issues/448) - Ignore désormais tout les noms trouvés après le mot-clé "destinataire(s)" dans le dernier tier du document lors de la recherche du médecin addresseur d'un courrier.
+- [#452](https://github.com/Refhi/Weda-Helper/issues/452) - Ignore désormais les lignes contenant plusieurs types d'examens (en général quand le centre d'imagerie annonce ses capacités d'examen).
+- [#464](https://github.com/Refhi/Weda-Helper/issues/464) - Prend en compte le décochage de l'option Titre Automatique (échanges sécurisés).
+- [#464](https://github.com/Refhi/Weda-Helper/issues/464) - Amélioration des messages d'erreurs lors de l'import des échanges sécurisés.
+
+
 # [2.13] - Gel des demandes de fonctionnalités
 ## AnnonceS : (les détails de la mise à jour sont après)
 Sommaire :